Fixed mp

+++ Projectiles/DimWarp2.cs
@@ -12,7 +12,6 @@ namespace TheConfectionRebirth.Projectiles
             Projectile.width = 32;
             Projectile.height = 38;
             Projectile.friendly = true;
-            Projectile.hostile = true;
             Projectile.tileCollide = true;
             Projectile.timeLeft = 10;
             Projectile.ai[0] = 0;
@@ -31,27 +30,40 @@ namespace TheConfectionRebirth.Projectiles
                 if (++Projectile.frame >= Main.projFrames[Projectile.type])
                     Projectile.frame = 0;
             }
+            
+            
+            Player test = Main.player[Projectile.owner];
+            float x1 = Projectile.position.X;
+            float y1 = Projectile.position.Y;
+            float x2 = Projectile.position.X + Projectile.width;
+            float y2 = Projectile.position.Y + Projectile.height;
+            if (test != null && test.active && test.position.X + test.width > x1 && test.position.X < x2 && test.position.Y + test.height > y1 && test.position.Y < y2)
+            {
+                OnHitPlayer(test)
+            }
         }
 
-        public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
+        const float warpRad = 144;
+        public override bool CanHitNPC(NPC target, int damage, float knockback, bool crit)
         {
-            if (target.boss == false)
+            Vector2 diff = target.Center - Projectile.Center;
+            float detectRad = warpRad + MathF.Max(target.width, target.height);
+            if(target.boss == false && diff.X > -detectRad && diff.X < detectRad && diff.Y > -detectRad && diff.Y < detectRad && diff.LengthSquared() < detectRad * detectRad)
             {
                 Projectile warppoint = Main.player[Projectile.owner].GetModPlayer<ConfectionPlayer>().DimensionalWarp;
                 target.position.X = warppoint.position.X;
                 target.position.Y = warppoint.position.Y;
-                target.HealEffect(1);
                 Projectile.ai[0] = 1;
                 Projectile.Kill();
             }
+            return false;
         }
-        public override void OnHitPlayer(Player target, int damage, bool crit)
+        void OnHitPlayer(Player target)
         {
-            if (Main.player[Main.myPlayer] == target)
-            {
+            //if (Main.player[Main.myPlayer] == target)
+            //{
                 Projectile warppoint = Main.player[Projectile.owner].GetModPlayer<ConfectionPlayer>().DimensionalWarp;
                 target.Teleport(warppoint.position, 1);
-                target.HealEffect(1);
                 if (target.HasBuff(ModContent.BuffType<Buffs.GoneBananas>()))
                 {
                     target.Hurt(PlayerDeathReason.ByCustomReason("DimensionSplit"), (int)((target.statLifeMax2 + target.statDefense) * (target.endurance + 1) / 7), 0);
@@ -59,7 +71,7 @@ namespace TheConfectionRebirth.Projectiles
                 target.AddBuff(ModContent.BuffType<Buffs.GoneBananas>(), 360);
                 Projectile.ai[0] = 1;
                 Projectile.Kill();
-            }
+            //}
         }
         public override bool PreKill(int timeLeft)
         {
